requirements.txt
aioredis==1.3.1
amqp==5.0.6
anyio==3.3.4
asgiref==3.4.1
async-timeout==4.0.1
autopep8==1.6.0
bidict==0.21.4
billiard==3.6.4.0
celery==5.2.0
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
fastapi==0.70.0
h11==0.12.0
hiredis==2.0.0
httptools==0.2.0
idna==3.3
kombu==5.2.1
prompt-toolkit==3.0.22
pycodestyle==2.8.0
pydantic==1.8.2
python-dotenv==0.19.1
python-engineio==4.3.0
python-socketio==5.4.1
pytz==2021.3
PyYAML==6.0
redis==3.5.3
six==1.16.0
sniffio==1.2.0
starlette==0.16.0
toml==0.10.2
typing-extensions==3.10.0.2
uvicorn==0.15.0
uvloop==0.16.0
vine==5.0.0
watchgod==0.7
wcwidth==0.2.5
websockets==10.0
